+++ routes/api.js
@@ -33,6 +33,13 @@ const yolovx = require("./training/yolovx");
 const uploadPreWeights = require("./training/uploadPreWeights");
 const yoloRun = require("./training/yoloRun");
 const deleteRun = require("./training/deleteRun");
+const python = require("./training/python");
+const darknet = require("./training/darknet");
+const removePath = require("./training/movePath");
+const removeDarknetPath = require("./training/removeDarknetPath");
+const removeWeights = require("./training/removeWeights");
+const removeScript = require("./training/removeScript");
+const run = require("./training/run");
 
 const createProject = require("./projects/createProject");
 const updateProject = require("./projects/updateProject");
@@ -48,6 +55,7 @@ const updateLabels = require("./labelling/updateLabels");
 
 const downloadDataset = require("./downloads/downloadDataset");
 const downloadProject = require("./downloads/downloadProject");
+const downloadScript = require("./downloads/downloadScript");
 
 const test = require("./tests/test");
 const mergeTest = require("./tests/mergeTest");
@@ -58,7 +66,7 @@ api.post("/login", login);
 api.post("/signup", signup);
 api.post("/addUser", addUser);
 
-// CLASSIFICATION ROUTES
+// TRAINING ROUTES
 api.post("/api/createC", createClassification);
 api.post("/addClasses", addClasses);
 api.post("/upload_weights", uploadWeights);
@@ -66,6 +74,13 @@ api.post("/yolovx", yolovx);
 api.post("/upload_pre_weights", uploadPreWeights);
 api.post("/yolo-run", yoloRun);
 api.post("/deleteRun", deleteRun);
+api.post("/python", python);
+api.post("/darknet", darknet);
+api.post("/remove_path", removePath);
+api.post("/remove_darknet_path", removeDarknetPath);
+api.post("/remove_weights", removeWeights);
+api.post("/remove_script", removeScript);
+api.post("/run", run);
 
 // PROJECT ROUTES
 api.post("/createP", createProject);
@@ -84,552 +99,13 @@ api.post("/updateLabels", updateLabels);
 // DOWNLOAD ROUTES
 api.post("/downloadDataset", downloadDataset);
 api.post("/downloadProject", downloadProject);
+api.post("/downloadScript", downloadScript);
 
 // TEST ROUTES
 api.post("/test", test);
 api.post("/mergeTest", mergeTest);
 
-api.post("/python", async (req, res) => {
-    console.log("new python path");
-
-    var PName = req.body.PName,
-        Admin = req.body.Admin,
-        user = req.cookies.Username,
-        python_path = req.body.python_path;
-
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/", // $LABELING_TOOL_PATH/public/projects/
-        project_path = main_path + Admin + "-" + PName, // $LABELING_TOOL_PATH/public/projects/Admin-project_name
-        images_path = project_path + "/images", // $LABELING_TOOL_PATH/public/projects/Admin-project_name/images
-        downloads_path = main_path + user + "_Downloads",
-        training_path = project_path + "/training",
-        python_path_file = training_path + "/Paths.txt";
-
-    console.log("python_path: ", python_path);
-    console.log("Exists: ", fs.existsSync(python_path));
-    if (fs.existsSync(python_path)) {
-        fs.writeFile(
-            python_path_file,
-            python_path + "\n",
-            { flag: "a" },
-            function (err) {
-                if (err) {
-                    console.log(err);
-                }
-            },
-        );
-
-        res.send({ Success: "Python Path Saved" });
-    } else {
-        res.send({ Success: `ERROR! ${python_path} is not a valid path` });
-    }
-});
-
-api.post("/darknet", async (req, res) => {
-    console.log("is this right pages");
-
-    var PName = req.body.PName,
-        Admin = req.body.Admin,
-        user = req.cookies.Username,
-        darknet_path = req.body.darknet_path;
-
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/", // $LABELING_TOOL_PATH/public/projects/
-        project_path = main_path + Admin + "-" + PName, // $LABELING_TOOL_PATH/public/projects/Admin-project_name
-        images_path = project_path + "/images", // $LABELING_TOOL_PATH/public/projects/Admin-project_name/images
-        downloads_path = main_path + user + "_Downloads",
-        training_path = project_path + "/training",
-        darknet_path_file = training_path + "/darknetPaths.txt";
-
-    console.log("darknet_path: ", darknet_path);
-    console.log("Exists: ", fs.existsSync(darknet_path));
-    if (fs.existsSync(darknet_path)) {
-        fs.writeFile(
-            darknet_path_file,
-            darknet_path + "\n",
-            { flag: "a" },
-            function (err) {
-                if (err) {
-                    console.log(err);
-                }
-            },
-        );
-
-        res.send({ Success: "Darknet Path Saved" });
-    } else {
-        res.send({ Success: `ERROR! ${darknet_path} is not a valid path` });
-    }
-});
-
-api.post("/remove_path", async (req, res) => {
-    console.log("remove python path");
-
-    var PName = req.body.PName,
-        Admin = req.body.Admin,
-        IDX = parseInt(req.body.IDX),
-        user = req.cookies.Username;
-
-    var remove_paths_arr = [];
-    remove_paths_arr.push(req.body["paths[]"]);
-    var remove_paths = [];
-    remove_paths = remove_paths.concat
-        .apply(remove_paths, remove_paths_arr)
-        .filter(Boolean);
-
-    console.log("remove_paths: ", remove_paths);
-
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/", // $LABELING_TOOL_PATH/public/projects/
-        project_path = main_path + Admin + "-" + PName, // $LABELING_TOOL_PATH/public/projects/Admin-project_name
-        images_path = project_path + "/images", // $LABELING_TOOL_PATH/public/projects/Admin-project_name/images
-        downloads_path = main_path + user + "_Downloads",
-        training_path = project_path + "/training",
-        python_path_file = training_path + "/Paths.txt";
-
-    var current_paths_arr = [];
-    current_paths_arr.push(
-        fs.readFileSync(python_path_file, "utf-8").split("\n").filter(Boolean),
-    );
-    var current_paths = [];
-    current_paths = current_paths.concat
-        .apply(current_paths, current_paths_arr)
-        .filter(Boolean);
-
-    var new_paths = "";
-    for (var i = 0; i < current_paths.length; i++) {
-        if (remove_paths.includes(current_paths[i])) {
-            continue;
-        }
-        new_paths = `${new_paths}${current_paths[i]}\n`;
-    }
-
-    fs.writeFile(python_path_file, new_paths, (err) => {
-        if (err) throw err;
-
-        return res.redirect("/config?IDX=" + IDX);
-    });
-});
-
-api.post("/remove_darknet_path", async (req, res) => {
-    console.log("remove darknet path");
-
-    var PName = req.body.PName,
-        Admin = req.body.Admin,
-        IDX = parseInt(req.body.IDX),
-        user = req.cookies.Username;
-
-    var remove_paths = req.body["darknet_paths[]"];
-    console.log("remove_paths: ", remove_paths);
-
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/", // $LABELING_TOOL_PATH/public/projects/
-        project_path = main_path + Admin + "-" + PName, // $LABELING_TOOL_PATH/public/projects/Admin-project_name
-        images_path = project_path + "/images", // $LABELING_TOOL_PATH/public/projects/Admin-project_name/images
-        downloads_path = main_path + user + "_Downloads",
-        training_path = project_path + "/training",
-        darknet_path_file = training_path + "/darknetPaths.txt",
-        python_path_file = training_path + "/Paths.txt";
-
-    var current_paths_arr = [];
-    current_paths_arr.push(
-        fs.readFileSync(darknet_path_file, "utf-8").split("\n").filter(Boolean),
-    );
-    var current_paths = [];
-    current_paths = current_paths.concat
-        .apply(current_paths, current_paths_arr)
-        .filter(Boolean);
-
-    var new_paths = "";
-
-    var darknet_path = new Set();
-
-    for (var i = 0; i < current_paths.length; i++) {
-        if (remove_paths.includes(current_paths[i])) {
-            darknet_path.add(current_paths[i]);
-            continue;
-        }
-        new_paths = `${new_paths}${current_paths[i]}\n`;
-    }
-
-    fs.writeFile(darknet_path_file, new_paths, (err) => {
-        if (err) throw err;
-
-        const drknt_temp = darknet_path.values();
-        for (var i = 0; i < darknet_path.size; i++) {
-            var current_darknet_path = drknt_temp.next().value;
-            var darknetFiles = readdirSync(current_darknet_path);
-            for (var i = 0; i < darknetFiles.length; i++) {
-                if (darknetFiles[i].split("-")[0] == user) {
-                    rimraf(
-                        current_darknet_path + "/" + darknetFiles[i],
-                        (err) => {
-                            if (err) {
-                                console.error(
-                                    "there was an error with the user contents: ",
-                                    err,
-                                );
-                            } else {
-                                console.log(
-                                    "Darknet user project contents successfuly deleted",
-                                );
-                            }
-                        },
-                    );
-                }
-            }
-        }
-        return res.redirect("/config?IDX=" + IDX);
-    });
-});
-
-api.post("/remove_weights", async (req, res) => {
-    console.log("remove weights file");
-
-    var PName = req.body.PName,
-        Admin = req.body.Admin,
-        IDX = parseInt(req.body.IDX),
-        user = req.cookies.Username,
-        weights_arr = [];
-    weights_arr.push(req.body["weights[]"]);
-    weights = [];
-    weights = weights.concat.apply(weights, weights_arr).filter(Boolean);
-
-    console.log("remove_weights weights: ", weights);
-    console.log("remove_weights weights.length: ", weights.length);
-
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/", // $LABELING_TOOL_PATH/public/projects/
-        project_path = main_path + Admin + "-" + PName, // $LABELING_TOOL_PATH/public/projects/Admin-project_name
-        images_path = project_path + "/images", // $LABELING_TOOL_PATH/public/projects/Admin-project_name/images
-        downloads_path = main_path + user + "_Downloads",
-        training_path = project_path + "/training",
-        script_path = training_path + "/python",
-        weights_path = training_path + "/weights",
-        python_path_file = training_path + "/Paths.txt";
-
-    for (var i = 0; i < weights.length; i++) {
-        weight = `${weights_path}/${weights[i]}`;
-
-        fs.unlink(weight, (error) => {
-            if (error) {
-                console.log(error);
-            }
-        });
-    }
-
-    return res.redirect("/config?IDX=" + IDX);
-});
-
-api.post("/remove_script", async (req, res) => {
-    console.log("remove python script");
-
-    var PName = req.body.PName,
-        Admin = req.body.Admin,
-        IDX = parseInt(req.body.IDX),
-        user = req.cookies.Username,
-        scripts = req.body["scripts[]"].split(",");
-
-    console.log("remove_script scripts: ", scripts);
-    console.log("remove_script scripts.length: ", scripts.length);
-
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/", // $LABELING_TOOL_PATH/public/projects/
-        project_path = main_path + Admin + "-" + PName, // $LABELING_TOOL_PATH/public/projects/Admin-project_name
-        images_path = project_path + "/images", // $LABELING_TOOL_PATH/public/projects/Admin-project_name/images
-        downloads_path = main_path + user + "_Downloads",
-        training_path = project_path + "/training",
-        script_path = training_path + "/python",
-        python_path_file = training_path + "/Paths.txt";
-
-    var new_paths = "";
-    for (var i = 0; i < scripts.length; i++) {
-        script = `${script_path}/${scripts[i]}`;
-        fs.unlink(script, (error) => {
-            console.log(error);
-        });
-    }
-
-    return res.redirect("/config?IDX=" + IDX);
-});
-
-api.post("/run", async (req, res) => {
-    console.log("run script");
-
-    const { exec } = require("child_process");
-
-    console.log("date: ", Date.now());
-    var date = Date.now();
-
-    var PName = req.body.PName,
-        Admin = req.body.Admin,
-        user = req.cookies.Username,
-        python_path = req.body.python_path,
-        script = req.body.script,
-        log = `${date}.log`,
-        weights = req.body.weights,
-        TrainingPercent = parseFloat(req.body.TrainingPercent);
-
-    if (TrainingPercent > 1) {
-        TrainingPercent = TrainingPercent / 100;
-    }
-    var options = "";
-
-    if (req.body.options) {
-        options = req.body.options;
-    } else {
-        options = "EMPTY";
-    }
-    console.log("options: ", options);
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/", // $LABELING_TOOL_PATH/public/projects/
-        project_path = main_path + Admin + "-" + PName, // $LABELING_TOOL_PATH/public/projects/project_name
-        images_path = project_path + "/images", // $LABELING_TOOL_PATH/public/projects/project_name/images
-        abs_images_path = path.join(__dirname, images_path),
-        downloads_path = main_path + user + "_Downloads",
-        training_path = project_path + "/training",
-        logs_path = training_path + "/logs",
-        run_path = `${logs_path}/${date}`,
-        weights_file = training_path + "/weights/" + weights,
-        python_script = training_path + "/python/" + script,
-        wrapper_path =
-            public_path + "controllers/training/train_data_from_project.py",
-        project_db = `${project_path}/${PName}.db`;
-
-    if (!fs.existsSync(run_path)) {
-        fs.mkdirSync(run_path);
-    }
-
-    fs.writeFile(`${run_path}/${log}`, "", (err) => {
-        if (err) throw err;
-    });
-
-    ///////////////////Connect to database /////////////////////////////////////////
-
-    var rundb = new sqlite3.Database(project_db, (err) => {
-        if (err) {
-            return console.error(err.message);
-        }
-        console.log("Connected to rundb.");
-    });
-    rundb.getAsync = function (sql) {
-        var that = this;
-        return new Promise(function (resolve, reject) {
-            that.get(sql, function (err, row) {
-                if (err) {
-                    console.log("runAsync ERROR! ", err);
-                    reject(err);
-                } else resolve(row);
-            });
-        }).catch((err) => {
-            console.error(err);
-        });
-    };
-    rundb.allAsync = function (sql) {
-        var that = this;
-        return new Promise(function (resolve, reject) {
-            that.all(sql, function (err, row) {
-                if (err) {
-                    console.log("runAsync ERROR! ", err);
-                    reject(err);
-                } else resolve(row);
-            });
-        }).catch((err) => {
-            console.error(err);
-        });
-    };
-    rundb.runAsync = function (sql) {
-        var that = this;
-        return new Promise(function (resolve, reject) {
-            that.run(sql, function (err, row) {
-                if (err) {
-                    console.log("runAsync ERROR! ", err);
-                    reject(err);
-                } else resolve(row);
-            });
-        }).catch((err) => {
-            console.error(err);
-        });
-    };
-
-    ///////////////Create Tensorflow training csv /////////////////////////////////
-
-    var labels = await rundb.allAsync("SELECT * FROM Labels");
-    var xmin = 0;
-    var xmax = 0;
-    var ymin = 0;
-    var ymax = 0;
-    var data = "filename,width,height,class,xmin,ymin,xmax,ymax\n";
-    var i = 0;
-    var image = "";
-    for (i = 0; i < Math.floor(labels.length * TrainingPercent); i++) {
-        image = path.join(abs_images_path, labels[i].IName);
-        xmin = labels[i].X;
-        xmax = xmin + labels[i].W;
-        ymin = labels[i].Y;
-        ymax = ymin + labels[i].Y;
-        data =
-            data +
-            image +
-            "," +
-            labels[i].W +
-            "," +
-            labels[i].H +
-            "," +
-            labels[i].CName +
-            "," +
-            xmin +
-            "," +
-            ymin +
-            "," +
-            xmax +
-            "," +
-            ymax +
-            "\n";
-    }
-    var training_csv = run_path + "/" + PName + "_train.csv";
-    fs.writeFile(training_csv, data, (err) => {
-        if (err) throw err;
-        console.log("done writing csv");
-    });
-
-    ///////////////Create Tensorflow Validate csv /////////////////////////////////
-    var data = "filename,width,height,class,xmin,ymin,xmax,ymax\n";
-    for (var j = i; j < labels.length; j++) {
-        image = path.join(abs_images_path, labels[j].IName);
-        xmin = labels[j].X;
-        xmax = xmin + labels[j].W;
-        ymin = labels[j].Y;
-        ymax = ymin + labels[j].Y;
-        data =
-            data +
-            image +
-            "," +
-            labels[j].W +
-            "," +
-            labels[j].H +
-            "," +
-            labels[j].CName +
-            "," +
-            xmin +
-            "," +
-            ymin +
-            "," +
-            xmax +
-            "," +
-            ymax +
-            "\n";
-    }
-    var validation_csv = run_path + "/" + PName + "_validate.csv";
-    fs.writeFile(validation_csv, data, (err) => {
-        if (err) throw err;
-        console.log("done writing csv");
-    });
-
-    rundb.close((err) => {
-        if (err) {
-            console.log(err);
-        } else {
-            console.log("rundb closed successfully");
-        }
-    });
-
-    //Create error file name
-    var err_file = `${date}-error.log`;
-
-    // Call Chris's script here
-    // Pass in python path, script, and options
-
-    //TODO: Add weights, training.csv, and validate.csv options to command
-    options = options.replace("<data_dir>", images_path);
-    options = options.replace("<training_csv>", training_csv);
-    options = options.replace("<validation_csv>", validation_csv);
-    options = options.replace("<output_dir>", run_path);
-    options = options.replace("<weights>", weights_file);
-
-    var cmd = `${wrapper_path} -p ${python_path} -s ${python_script} -l ${run_path}/${log} -o "${options}"`;
-    var success = "Training complete";
-    var error = "";
-    process.chdir(run_path);
-
-    var child = exec(cmd, (err, stdout, stderr) => {
-        if (err) {
-            console.log(`This is the error: ${err.message}`);
-            success = err.message;
-            fs.writeFile(`${run_path}/${err_file}`, success, (err) => {
-                if (err) throw err;
-            });
-        } else if (stderr) {
-            console.log(`This is the stderr: ${stderr}`);
-            fs.writeFile(`${run_path}/${err_file}`, stderr, (err) => {
-                if (err) throw err;
-            });
-            //return;
-        }
-        console.log("stdout: ", stdout);
-        console.log("stderr: ", stderr);
-        console.log("err: ", err);
-        console.log("The script has finished running");
-        fs.writeFile(`${run_path}/done.log`, success, (err) => {
-            if (err) throw err;
-        });
-    });
-
-    res.send({ Success: `Training Started` });
-    process.chdir(`${public_path}routes`);
-});
-
 ////////////////////////////Download page/////////////////////////////////////////////////
-
-api.post("/downloadScript", async (req, res) => {
-    console.log("downloadingScripts");
-
-    // get URL variables
-    var PName = req.body.PName,
-        admin = req.body.Admin,
-        IDX = parseInt(req.body.IDX),
-        // mult = parseInt(req.body.mult),
-        user = req.cookies.Username;
-    //scripts = req.body["scripts[]"];
-
-    var scripts_arr = [];
-    scripts_arr.push(req.body["scripts[]"]);
-    var scripts = [];
-    scripts = scripts.concat.apply(scripts, scripts_arr).filter(Boolean);
-
-    console.log("scritps: ", scripts);
-    // Set paths
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/", // $LABELING_TOOL_PATH/public/projects/
-        project_path = main_path + admin + "-" + PName, // $LABELING_TOOL_PATH/public/projects/project_name
-        images_path = project_path + "/images", // $LABELING_TOOL_PATH/public/projects/project_name/images
-        downloads_path = main_path + user + "_Downloads",
-        training_path = project_path + "/training",
-        python_path = training_path + "/python",
-        logs_path = training_path + "/logs";
-
-    // Create zipfile
-
-    var output = fs.createWriteStream(downloads_path + "/scripts.zip");
-    var archive = archiver("zip");
-
-    output.on("close", function () {
-        console.log(archive.pointer() + " total bytes");
-        console.log(
-            "archiver has been finalized and the output file descriptor has closed.",
-        );
-        res.download(downloads_path + "/scripts.zip");
-    });
-    archive.on("error", function (err) {
-        throw err;
-    });
-
-    archive.pipe(output);
-    for (var i = 0; i < scripts.length; i++) {
-        script = `${python_path}/${scripts[i]}`;
-        archive.file(script, { name: scripts[i] });
-    }
-    archive.finalize();
-});
 
 api.post("/downloadWeights", async (req, res) => {
     console.log("downloadingWeights");
