+++ routes/training/darknet.js
@@ -1,10 +1,10 @@
-async function python(req, res) {
-    console.log("new python path");
+async function darknet(req, res) {
+    console.log("is this right pages");
 
     var PName = req.body.PName,
         Admin = req.body.Admin,
         user = req.cookies.Username,
-        python_path = req.body.python_path;
+        darknet_path = req.body.darknet_path;
 
     var public_path = __dirname.replace("routes", "").replace("training", ""),
         main_path = public_path + "public/projects/", // $LABELING_TOOL_PATH/public/projects/
@@ -12,14 +12,14 @@ async function python(req, res) {
         images_path = project_path + "/images", // $LABELING_TOOL_PATH/public/projects/Admin-project_name/images
         downloads_path = main_path + user + "_Downloads",
         training_path = project_path + "/training",
-        python_path_file = training_path + "/Paths.txt";
+        darknet_path_file = training_path + "/darknetPaths.txt";
 
-    console.log("python_path: ", python_path);
-    console.log("Exists: ", fs.existsSync(python_path));
-    if (fs.existsSync(python_path)) {
+    console.log("darknet_path: ", darknet_path);
+    console.log("Exists: ", fs.existsSync(darknet_path));
+    if (fs.existsSync(darknet_path)) {
         fs.writeFile(
-            python_path_file,
-            python_path + "\n",
+            darknet_path_file,
+            darknet_path + "\n",
             { flag: "a" },
             function (err) {
                 if (err) {
@@ -28,10 +28,10 @@ async function python(req, res) {
             },
         );
 
-        res.send({ Success: "Python Path Saved" });
+        res.send({ Success: "Darknet Path Saved" });
     } else {
-        res.send({ Success: `ERROR! ${python_path} is not a valid path` });
+        res.send({ Success: `ERROR! ${darknet_path} is not a valid path` });
     }
 }
 
-module.exports = python;
+module.exports = darknet;
